Modified mouseAim.cs
Turning the ship is now done on a continuously basis rather than an
instantaneous basis. You basically point the mouse away from the middle
of the screen to turn the ship around. Currently not inverted along
y-axis. Watch out for deadlocking at the 90 degree angles of the y-axis
of ship.